requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
pydantic==2.7.1
pillow==10.3.0
google-generativeai
python-dotenv
